Fix: GridSize coercion type mismatch
Changed MinGridSize and MaxGridSize from double to int to match GridSize property type.
This fixes the ArgumentException: '10' ist kein gültiger Wert für die Eigenschaft 'GridSize'.

The CoerceGridSize method was returning double values from Math.Max/Min,
but GridSize DependencyProperty expects int values.

+++ src/DigitalSignage.Server/Controls/ModernDesignerCanvas.cs
@@ -55,8 +55,8 @@ public class ModernDesignerCanvas : Canvas
 
     private const double RulerTickSize = 12.0;
     private const string ElementTypeDataFormat = "DesignerElementType";
-    private const double MinGridSize = 5.0;
-    private const double MaxGridSize = 100.0;
+    private const int MinGridSize = 5;
+    private const int MaxGridSize = 100;
 
     #endregion
